backend/pyproject.toml
[project]
name = "pitch-maker-backend"
version = "0.1.0"
description = "FastAPI backend for Pitch Maker"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0,<1.0.0",
  "uvicorn[standard]>=0.32.0,<1.0.0",
  "httpx>=0.28.0,<1.0.0",
  "mistralai>=1.2.0,<2.0.0",
  "elevenlabs>=2.7.0,<3.0.0",
  "fal-client>=0.5.6,<1.0.0",
  "pydantic-settings>=2.6.1,<3.0.0",
  "python-dotenv>=1.0.1,<2.0.0",
  "pillow>=11.0.0,<12.0.0",
]

[dependency-groups]
dev = [
  "pytest>=8.3.3,<9.0.0",
  "pytest-asyncio>=0.24.0,<1.0.0",
  "ruff>=0.8.0,<1.0.0",
  "mypy>=1.13.0,<2.0.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
